requirements.txt
streamlit
plotly.express
simple-salesforce
pandas
python-dotenv